Use committed WikidataItem.php file
The previously referenced path –

/data/project/magnustools/public_html/php/WikidataItem.php

– only exists on Toolforge: not just due to the absolute /data/project
component, but also because public_html/php/WikidataItem.php is not
committed in this Git repository. Use classes/WikidataItem.php instead,
which is almost identical except that the class is called
Toolforge\WikidataItem instead of \WDI.

This effectively resurrects commit c59a765580, which made use of the
relative path but did not adjust the class name.

Co-Authored-By: laserjeyes <johannes.hentschel@epfl.ch>
Signed-Off-By: Itamar Givon <itamar.givon@wikimedia.de>